CAPP-10694 Make user attribute contain push id and client as well as alias

+++ jaxrs/src/main/java/org/jboss/aerogear/unifiedpush/rest/registry/applications/AliasEndpoint.java
@@ -603,7 +603,7 @@ public class AliasEndpoint extends AbstractBaseEndpoint {
 	@ReturnType("org.jboss.aerogear.unifiedpush.api.Alias")
 	public Response updateKcGuids(@Context HttpServletRequest request) {
 		try {
-			long updated = aliasService.updateKCUsersGuids();
+			int updated = aliasService.updateKCUsersGuids();
 			return appendAllowOriginHeader(Response.ok(Collections.singletonMap("count", updated)), request);
 		} catch (Exception e) {
 			logger.error("Cannot update guids", e);
